+++ CHANGELOG.md
@@ -7,6 +7,8 @@ v0.2.0
   It's no longer `{.pure.}` and initial
   letter of all properties are lowercase
 * Fix `lookupStrict` for nim 0.17.3
+* `EProps` was renamed to `Prop` and
+  it's no longer `{.pure.}`
 
 v0.1.0
 ==================